premiere encapsulation pour l'affichage de texte 2D
